requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.0
validators>=0.22.0
pathvalidate>=3.2.0
python-magic>=0.4.27
chardet>=5.2.0
scipy>=1.11.0
seaborn>=0.12.0
matplotlib>=3.7.0
plotly>=5.15.0
scikit-learn>=1.3.0
nltk>=3.8.0
joblib>=1.3.0
pyarrow>=12.0.0
streamlit>=1.28.0

# Optional — install separately as needed:
xgboost>=2.0.0
lightgbm>=4.0.0
Pillow>=10.0.0
librosa>=0.10.0
kaggle>=1.5.0
gdown>=4.7.0
